moved extensions and hooks to separate folders
